Separerte ut kryptering i egen klasse, fikset test
